fix: handle boolean field properly
- update field extract field for boolean data type to allow false as valid value
- update unit tests
- add boolean unit test
- update test data for local development
