+++ functions/__tests__/__mocks__/console.ts
@@ -1,4 +1,4 @@
-import { logger } from "firebase-functions";
+import { logger } from 'firebase-functions';
 
 export const mockConsoleInfo = jest.spyOn(logger, "info").mockImplementation();
