+++ functions/__tests__/data/document.ts
@@ -55,5 +55,6 @@ export default {
     "Drama",
     "Crime"
   ],
+  'popular': false,
   "objectID": "439817390"
 }
