+++ functions/__tests__/functions.test.ts
@@ -1,11 +1,8 @@
 import algoliasearch from 'algoliasearch';
-import { firestore } from 'firebase-admin/lib/firestore';
 import * as functionsTestInit from 'firebase-functions-test';
 import mockedEnv from 'mocked-env';
 import { mocked } from 'ts-jest/utils';
-import { mockConsoleInfo } from './__mocks__/console';
-import {version} from '../src/version';
-import testDocument, { testReleaseDate } from './data/document';
+import { version } from '../src/version';
 
 jest.mock('algoliasearch');
