+++ functions/__tests__/functions.update.undefined.fields.test.ts
@@ -14,7 +14,7 @@ const defaultEnvironment = {
   ALGOLIA_API_KEY: '********',
   ALGOLIA_INDEX_NAME: 'algolia-index-name',
   COLLECTION_PATH: 'movies',
-  FIELDS: 'title,awards,meta'
+  FIELDS: ''
 };
 
 export const mockExport = (document, data) => {
@@ -67,9 +67,15 @@ describe('extension', () => {
       functionsConfig = config;
     });
 
-    test('functions runs with a create', async () => {
-      const beforeSnapshot = functionsTest.firestore.makeDocumentSnapshot({}, 'document/1');
-      const afterSnapshot = functionsTest.firestore.makeDocumentSnapshot(testDocument, 'document/1');
+    test('functions runs with an update on non indexable field', async () => {
+      restoreEnv = mockedEnv(defaultEnvironment);
+      config = require('../src/config').default;
+
+      const beforeSnapshot = functionsTest.firestore.makeDocumentSnapshot(testDocument, 'document/1');
+      const afterSnapshot = functionsTest.firestore.makeDocumentSnapshot({
+        ...testDocument,
+        title: 'The Prison'
+      }, 'document/1');
 
       const documentChange = functionsTest.makeChange(
         beforeSnapshot,
@@ -86,6 +92,7 @@ describe('extension', () => {
         functionsConfig
       );
       const payload = {
+        ...testDocument,
         'objectID': afterSnapshot.id,
         'title': afterSnapshot.data().title,
         'awards': [
@@ -100,11 +107,56 @@ describe('extension', () => {
         }
       }
       expect(mockConsoleInfo).toBeCalledWith(
-        `Creating new Algolia index for document ${ afterSnapshot.id }`,
+        `Updating existing Algolia index for document ${ afterSnapshot.id }`,
         payload
       );
+      expect(mockedPartialUpdateObject).toBeCalledWith(payload, { createIfNotExists: true });
+    });
 
-      expect(mockedPartialUpdateObject).toBeCalledWith(payload,  { createIfNotExists: true });
+    test('functions runs with an update with falsy values', async () => {
+      const beforeSnapshot = functionsTest.firestore.makeDocumentSnapshot(testDocument, 'document/1');
+      const afterSnapshot = functionsTest.firestore.makeDocumentSnapshot({
+        ...testDocument,
+        popular: false,
+        rating: 0
+      }, 'document/1');
+
+      const documentChange = functionsTest.makeChange(
+        beforeSnapshot,
+        afterSnapshot
+      );
+
+      const data = {};
+      const callResult = await mockExport(documentChange, data);
+
+      expect(callResult).toBeUndefined();
+      expect(mockConsoleInfo).toBeCalledTimes(2);
+      expect(mockConsoleInfo).toBeCalledWith(
+        'Started extension execution with configuration',
+        functionsConfig
+      );
+      const payload = {
+        ...testDocument,
+        'objectID': afterSnapshot.id,
+        'title': afterSnapshot.data().title,
+        'awards': [
+          'awards/1'
+        ],
+        'meta': {
+          'releaseDate': testReleaseDate.getTime()
+        },
+        'popular': false,
+        'rating': 0,
+        'lastmodified': {
+          '_operation': 'IncrementSet',
+          'value': expect.any(Number)
+        }
+      }
+      expect(mockConsoleInfo).toBeCalledWith(
+        `Updating existing Algolia index for document ${ afterSnapshot.id }`,
+        payload
+      );
+      expect(mockedPartialUpdateObject).toBeCalledWith(payload, { createIfNotExists: true });
     });
   });
 });
